Adding travis and appveyor cotinuous integration tests
This is an automated pull request that includes test files for travis and appveyor.
Additionally you'll need to create an account on travis.org and appveyor.com and enable
the tests for this addon there.

Once that's done and this PR is merged any new commit to this addon will be tested against
the latest version of openFrameworks for all the supported platforms.